Move legend to new file
